Preliminary work for third-party plugins support
This rewrite is intended to be used for non-compliant third-party plugins that may have to modify the products, discounts, fees, and shipping arrays outside what is currently being processed in the respective functions.

For example, certain discounts, while processed during the discounts function, behave similar to products, and should therefore be treated like such. However, these should also be reported as discounts.