Use order_key instead of merchant_reference
The order id is now passed around in the return and callback URLs  in the 'key' parameter as an order_key rather than the 'merchant_reference' parameter which we use to identify the WooCommerce order.

We chose to move away for the 'merchant_reference' since Dintero does not permit modifying the merchant reference once it has been set once.