Add a workaround for misconfigured variant products
If the attribute value is not set for each of the variants, the default one will be used in all variants. On the front-end it will appear as a unique variant, but the product id, and variation id will be the same for all the variants. Since WooCommerce will create a variant for each of the attribute values, we can use the attribute value which the customer selected on the product page to uniquely identify the variant.

This should fix the problem with line_id not being unique when an order contain variants.