Unset the WC session on cancel payment in Express
When the customer cancels the payment, we'll unset the WC session, and then redirect to the checkout page.

Previously, we would check when the checkout was being rendered if the session is canceled at Dintero. This new approach is more general, and simply assumes that if the WC session for Dintero is not set, we'll automatically create a new session.