Use N/A as full name if not set on finalization
We already did this, but this was happening to early: as soon as the page loaded in. Thus the customer had to clear the fields, and then enter valid data.

With this change, we wait until we're about to finalize the purchase before we check if we need to fill the default value.