Clear the placeholder shipping address on failure
On Dintero Checkout + Embedded, the ship to a different address checkbox is checked by default, but the shipping address form is hidden (probably a bug in WC). We should not check this box unless we're sure we have shipping address data. Dintero will always provide the shipping country. If that's the only available data then no shipping address data is available.

Additionally, the placeholder data is still necessary for Dintero Checkout Express + Embedded when purchasing as a guest business customer. The placeholder for shipping last name has been replaced "blank braille pattern" to reduce the clutter.