Save billing address to shipping address fields when defaulting to shipping
If the default shipping destination is set to shipping, Woo will check the "ship to different address" checkbox, requiring the shipping address. However, a separate shipping address is only set if the merchant has enabled "Allow separate shipping address" in the Dintero plugin settings.

In the case of when this plugin setting is not enabled, we'll save the billing address to the shipping address fields.