Do not call save() when processing order item
Calling save() causes the order ID to return 0 (or sometimes `null`) in the `woocommerce_new_order_item` action hook if the priority is too low. This has caused a compatibility issue with plugins that rely on the order ID to be available.

Since the save() is not necessary when processing order items (it is however required for orders), we can safely remove it.

Furthermore, we don't need to call `update_meta_data` since the order line is only added once. Instead `add_meta_data` suffices.

Tested w/wo HPOS.