Get the order from the database before checking paid date
This is done to add a extra check in cases where the order might have had the redirect triggered twice causing the confirmation to run twice at the same time. By getting the order again from the database right before we check the date paid, it could prevent the order from being processed twice. This should be a minimal performance impact due to the cashing the WordPress database does inherently.